Updated to use static files
The CDN metroUI provided was not a direct version file, but instead the most recent version released. This causes WFDesk to break on updated versions of MetroUI.

A previous version of MetroUI was downloaded and then placed in this project to allow it to work again.

The same does not need to be done with jQuery, as it's a static versioned link.